requirements.txt
altgraph==0.17.3
attrs==23.1.0
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
configparse==0.1.5
decorator==4.4.2
Eel==0.16.0
einops==0.6.1
exceptiongroup==1.1.3
filelock==3.9.0
fsspec==2023.9.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
MarkupSafe==2.1.2
moviepy==1.0.3
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
pefile==2023.2.7
Pillow==9.3.0
proglog==0.1.10
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PySocks==1.7.1
PyWavelets==1.4.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qimage2ndarray==1.10.0
requests==2.28.1
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.8.30
timm==0.9.7
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.4.0
urllib3==1.26.13
whichcraft==0.6.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0
